Changed from float to double. Fixed non-windows compliant errors.

+++ BrickCounting/Brick.cpp
@@ -9,8 +9,8 @@ b: original brick placed in space.
 rb: rb in same configuration as b where b is original brick.
  */
 Brick::Brick(const Brick& b, const RectilinearBrick& rb) : angle(b.angle), level(b.level + rb.level()) {
-  float sina = sin(angle);
-  float cosa = cos(angle);
+  double sina = sin(angle);
+  double cosa = cos(angle);
   center.X = rb.x*cosa - rb.y*sina;
   center.Y = rb.x*sina + rb.y*cosa;  
 
@@ -24,13 +24,13 @@ Brick::Brick(const Brick& b, const RectilinearBrick& rb) : angle(b.angle), level
 /*
  Main constructor: Built from a RectilinearConfiguration by connecting to a brick at an angle.
  */
-Brick::Brick(const RectilinearBrick& b, const ConnectionPoint& p, const Point &origin, float originAngle, int8_t originLv) : center(b.x-p.x(), b.y-p.y()), angle(originAngle), level(b.level()+originLv) {  
+Brick::Brick(const RectilinearBrick& b, const ConnectionPoint& p, const Point &origin, double originAngle, int8_t originLv) : center(b.x-p.x(), b.y-p.y()), angle(originAngle), level(b.level()+originLv) {  
   std::cout << "Building brick. RB=" << b << std::endl << " center compared to connection: " << center.X << "," << center.Y << std::endl << " point to connect to: " << origin.X << "," << origin.Y << ", angle of that point: " << originAngle << std::endl;
   // center is now on b. Move by turn angle, then translate to origin:
   // Rotate:
-  float sina = sin(angle);
-  float cosa = cos(angle);
-  float oldX = center.X;
+  double sina = sin(angle);
+  double cosa = cos(angle);
+  double oldX = center.X;
   center.X = center.X*cosa - center.Y*sina;
   center.Y = oldX*sina + center.Y*cosa;
 
@@ -46,12 +46,12 @@ Brick::Brick(const RectilinearBrick& b, const ConnectionPoint& p, const Point &o
 }
 
 void Brick::toLDR(std::ofstream &os, int xx, int yy, int ldrColor) const {
-  float x = (this->center.X)*20;
-  float y = (this->center.Y)*20;
+  double x = (this->center.X)*20;
+  double y = (this->center.Y)*20;
   int z = -24*level;
 
-  float sina = sin(-angle+M_PI/2);
-  float cosa = cos(-angle+M_PI/2);
+  double sina = sin(-angle+M_PI/2);
+  double cosa = cos(-angle+M_PI/2);
 
   // Brick:
   os << std::fixed;
@@ -101,19 +101,19 @@ void Brick::moveBrickSoThisIsAxisAlignedAtOrigin(Brick &b) const {
   b.center.X = b.center.X-center.X;
   b.center.Y = b.center.Y-center.Y;
   // Rotate to make this->angle = 0.
-  float sina = sin(-angle);
-  float cosa = cos(-angle);
+  double sina = sin(-angle);
+  double cosa = cos(-angle);
   b.center.X = b.center.X*cosa - b.center.Y*sina;
   b.center.Y = b.center.X*sina + b.center.Y*cosa;
   b.angle = b.angle-angle;
 }
 
 void Brick::getBoxPOIs(Point *pois) const {
-  float sina = sin(angle);
-  float cosa = cos(angle);
+  double sina = sin(angle);
+  double cosa = cos(angle);
   // 4 corners:
-  float dx = VERTICAL_BRICK_CENTER_TO_SIDE;
-  float dy = VERTICAL_BRICK_CENTER_TO_TOP;
+  double dx = VERTICAL_BRICK_CENTER_TO_SIDE;
+  double dy = VERTICAL_BRICK_CENTER_TO_TOP;
   pois[0] = Point(center.X+(dx*cosa-dy*sina),  center.Y+(dx*sina+dy*cosa));
   pois[1] = Point(center.X+(-dx*cosa-dy*sina), center.Y+(-dx*sina+dy*cosa));
   pois[2] = Point(center.X+(dx*cosa+dy*sina),  center.Y+(dx*sina-dy*cosa));
@@ -124,9 +124,9 @@ void Brick::getBoxPOIs(Point *pois) const {
 }
 
 Point Brick::getStudPosition(ConnectionPointType type) const {
-  float sina = sin(angle);
-  float cosa = cos(angle);
-  float dx, dy;
+  double sina = sin(angle);
+  double cosa = cos(angle);
+  double dx, dy;
   switch(type) {
   case NW: dx = -HALF_STUD_DISTANCE; dy = STUD_AND_A_HALF_DISTANCE; break;
   case NE: dx = HALF_STUD_DISTANCE; dy = STUD_AND_A_HALF_DISTANCE; break;
@@ -134,17 +134,17 @@ Point Brick::getStudPosition(ConnectionPointType type) const {
   case SW: dx = -HALF_STUD_DISTANCE; dy = -STUD_AND_A_HALF_DISTANCE; break;
   default: dx = 0; dy = 0; std::cerr << "ARRR!" << std::endl; break;
   }
-  float dx2 = dx*cosa - dy*sina;
-  float dy2 = dx*sina + dy*cosa;
+  double dx2 = dx*cosa - dy*sina;
+  double dy2 = dx*sina + dy*cosa;
   return Point(center.X+dx2, center.Y+dy2);
 }
 
 void Brick::getStudPositions(Point *studs) const {
-  float sina = sin(angle);
-  float cosa = cos(angle);
+  double sina = sin(angle);
+  double cosa = cos(angle);
   // 4 inner:
-  float dx = HALF_STUD_DISTANCE;
-  float dy = HALF_STUD_DISTANCE;
+  double dx = HALF_STUD_DISTANCE;
+  double dy = HALF_STUD_DISTANCE;
   studs[0] = Point(center.X+(-dx*cosa-dy*sina), center.Y+(-dx*sina+dy*cosa));
   studs[1] = Point(center.X+(dx*cosa-dy*sina),  center.Y+(dx*sina+dy*cosa));
   studs[2] = Point(center.X+(dx*cosa+dy*sina),  center.Y+(dx*sina-dy*cosa));
@@ -197,8 +197,8 @@ bool Brick::boxesIntersect(const Brick &b) const {
       stud.Y = -stud.Y;
     if(stud.X < VERTICAL_BRICK_CENTER_TO_SIDE && stud.Y < VERTICAL_BRICK_CENTER_TO_TOP) {
       // Might intersect - check corner case:
-      const float cornerX = VERTICAL_BRICK_CENTER_TO_SIDE-STUD_RADIUS;
-      const float cornerY = VERTICAL_BRICK_CENTER_TO_TOP-STUD_RADIUS;
+      const double cornerX = VERTICAL_BRICK_CENTER_TO_SIDE-STUD_RADIUS;
+      const double cornerY = VERTICAL_BRICK_CENTER_TO_TOP-STUD_RADIUS;
       if(stud.X > cornerX && stud.Y > cornerY) {
 	return STUD_RADIUS*STUD_RADIUS > (stud.X-cornerX)*(stud.X-cornerX)+(stud.Y-cornerY)*(stud.Y-cornerY);
       }
@@ -215,8 +215,8 @@ bool Brick::boxesIntersect(const Brick &b) const {
 
     if(stud.X < VERTICAL_BRICK_CENTER_TO_SIDE && stud.Y < VERTICAL_BRICK_CENTER_TO_TOP) {
       // X check if it hits stud:
-      const float studX = STUD_DISTANCE;
-      const float studY = HALF_STUD_DISTANCE;    
+      const double studX = STUD_DISTANCE;
+      const double studY = HALF_STUD_DISTANCE;    
       if(SNAP_DISTANCE*SNAP_DISTANCE >= (stud.X-studX)*(stud.X-studX)+(stud.Y-studY)*(stud.Y-studY)) {
 	// We are already connected:
 	if(connected) {
@@ -237,8 +237,8 @@ bool Brick::boxesIntersect(const Brick &b) const {
       }      
 
       // Might intersect - check corner case:
-      const float cornerX = VERTICAL_BRICK_CENTER_TO_SIDE-STUD_RADIUS;
-      const float cornerY = VERTICAL_BRICK_CENTER_TO_TOP-STUD_RADIUS;
+      const double cornerX = VERTICAL_BRICK_CENTER_TO_SIDE-STUD_RADIUS;
+      const double cornerY = VERTICAL_BRICK_CENTER_TO_TOP-STUD_RADIUS;
       if(stud.X > cornerX && stud.Y > cornerY) {
 	return STUD_RADIUS*STUD_RADIUS > (stud.X-cornerX)*(stud.X-cornerX)+(stud.Y-cornerY)*(stud.Y-cornerY);
       }
